move to src/generated/src
