+++ sdk/keyvault/keyvault-keys/src/index.ts
@@ -6,9 +6,9 @@ import type { TokenCredential } from "@azure/core-auth";
 import { logger } from "./log.js";
 import { PageSettings, PagedAsyncIterableIterator } from "@azure/core-paging";
 import { PollOperationState, PollerLike } from "@azure/core-lro";
-import { DeletionRecoveryLevel, KnownDeletionRecoveryLevel } from "./generated/models/index.js";
-import type { KeyVaultClientOptionalParams } from "./generated/keyVaultClient.js";
-import { KeyVaultClient } from "./generated/keyVaultClient.js";
+import { DeletionRecoveryLevel, KnownDeletionRecoveryLevel } from "./generated/src/models/index.js";
+import type { KeyVaultClientOptionalParams } from "./generated/src/keyVaultClient.js";
+import { KeyVaultClient } from "./generated/src/keyVaultClient.js";
 import { SDK_VERSION } from "./constants.js";
 import { keyVaultAuthenticationPolicy } from "@azure/keyvault-common";
 import { DeleteKeyPoller } from "./lro/delete/poller.js";
