+++ sdk/keyvault/keyvault-keys/src/keysModels.ts
@@ -4,11 +4,11 @@
 import type * as coreClient from "@azure-rest/core-client";
 import type { ExtendedCommonClientOptions } from "@azure/core-http-compat";
 
-import type { DeletionRecoveryLevel } from "./generated/models/index.js";
+import type { DeletionRecoveryLevel } from "./generated/src/models/index.js";
 import {
   JsonWebKeyOperation as KeyOperation,
   JsonWebKeyType as KeyType,
-} from "./generated/models/index.js";
+} from "./generated/src/models/index.js";
 
 import type { KeyCurveName } from "./cryptographyClientModels.js";
