+++ sdk/keyvault/keyvault-keys/src/lro/keyVaultKeyPoller.ts
@@ -5,7 +5,7 @@ import type { OperationOptions } from "@azure-rest/core-client";
 import { delay } from "@azure/core-util";
 import type { PollOperation, PollOperationState } from "@azure/core-lro";
 import { Poller } from "@azure/core-lro";
-import type { KeyVaultClient } from "../generated/keyVaultClient.js";
+import type { KeyVaultClient } from "../generated/src/keyVaultClient.js";
 
 /**
  * Common parameters to a Key Vault Key Poller.
